* remove the /Wp64 switch * set the warning level to 4 (/W4) * fix some compiler warnings (unreferenced variables, missing typecasts)
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@11510 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/LogCache/CacheFileInBuffer.cpp
@@ -65,7 +65,7 @@ void CCacheFileInBuffer::ReadStreamOffsets()
 
 	// consistency check
 
-	if (lastStream - GetBuffer() != contentEnd)
+	if ((size_t)(lastStream - GetBuffer()) != contentEnd)
 		throw std::exception ("stream directory corrupted");
 }
